… datafusion/core/src/datasource/listing/table.rs
… datafusion/core/src/datasource/listing_table_factory.rs
… datafusion/core/src/datasource/physical_plan/mod.rs
… datafusion/core/src/physical_planner.rs
… datafusion/core/tests/fifo.rs