## Which issue does this PR close?

Part of #7079

## Rationale for this change

Serialization of "stateless" file types (where the serialized bytes of each record batch has no dependency on the serialized bytes of any other record batch), can be parallelized efficiently across all available CPU cores for a significant decrease in the time needed to write out the file. 

There is a tradeoff between write speed and memory utilization. If the ObjectStore writer cannot keep up with the data being serialized, bytes could accumulate in memory. ObjectStore puts are concurrent but not parallelized so the risk of higher memory usage increases as the number of cores in the system increases. We could potentially bound this increase in memory by throttling serialization tasks when more than X RecordBatches have been serialized but not yet written to the ObjectStore (**this is done!**).

## What changes are included in this PR?

Spawn a tokio task to serialize each each record batch in parallel for "stateless" file types, concurrently write the serialized bytes to an ObjectStore. 

## Benchmarking

To benchmark the performance difference in this PR, the following script is used. 
```bash
cd datafusion/benchmarks
./bench.sh data tpch10
```

```rust
use chrono;
use datafusion::prelude::*;
use datafusion_common::DataFusionError;
use object_store::local::LocalFileSystem;
use std::{io::Error, sync::Arc, time::Instant};
use url::Url;

const FILENAME: &str =
    "/home/dev/arrow-datafusion/benchmarks/data/tpch_sf10/lineitem/part-0.parquet";

#[tokio::main]
async fn main() -> Result<(), DataFusionError> {
    let _ctx = SessionContext::new();
    let local = Arc::new(LocalFileSystem::new());
    let local_url = Url::parse("file://local").unwrap();
    _ctx.runtime_env().register_object_store(&local_url, local);

    let _read_options = ParquetReadOptions::default();

    let start = Instant::now();
    let _df = _ctx
        .read_parquet(FILENAME, _read_options)
        .await
        .unwrap()
        //select a few columns with types compatible with write_json method
        .select_columns(&["l_orderkey", "l_partkey", "l_receiptdate"])?
        .repartition(Partitioning::Hash(vec![col("l_orderkey")], 4))?
        .cache()
        .await?;
    let elapsed = Instant::now() - start;
    println!("read parquet to memory took -> {elapsed:?}");

    let start3 = Instant::now();
    _df.clone()
        .write_csv("file://local/home/dev/arrow-datafusion/test_out/")
        .await?;
    let elapsed3 = Instant::now() - start3;
    println!("write as csv to disk took -> {elapsed3:?}");

    let start2 = Instant::now();
    _df.clone()
        .write_json("file://local/home/dev/arrow-datafusion/test_out/")
        .await?;
    let elapsed2 = Instant::now() - start2;
    println!("write as json to disk took -> {elapsed2:?}");

    Ok(())
}

```

Tables below report results on a 16c/32t development server with 128Gb of memory writing to fast local nvme storage. This is an ideal case, scaling will not be as good on systems with fewer threads, significantly less memory, or slower/remote storage.

### Execution Time in Seconds
  | main | This PR | % diff
-- | -- | -- | --
1 JSON file | 220.31 | 20.76 | -91%
1 CSV file | 144.31 | 11.82 | -92%
4 JSON files | 224.14 | 19.05 | -92%
4 CSV files | 147.11 | 11.46 | -92%

### Peak Memory Usage in GB
  | main | This PR | % diff
-- | -- | -- | --
1 JSON file | 0.55 | 3.9 | 609%
1 CSV file | 0.55 | 0.85 | 55%
4 JSON files | 0.65 | 3.9 | 500%
4 CSV files | 0.65 | 1.00 | 54%

After incorporating the channel method written by @metesynnada to introduce backpressure, memory usage increase is consistently ~200MB is all tests with no noticeable drop in performance.

Note: Peak memory usage figure subtracts the peak memory usage required to cache the DataFrame and not perform any writes.

On a laptop with 15t, I see 28.42s and 42.94s for the 4CSV and 4JSON file tests respectively. 

Without the caching step, each write only takes about 1 extra second.

## What about Parquet?

Serializing a parquet file is more complex (the parquet writer maintains internal state across record batches).. Serializing two different parquet files in parallel is trivial, but parallelizing serialization of a single large parquet file is more difficult. It may be possible by constructing multiple row groups in parallel. I intend to work on this in another PR.

## Are these changes tested?

Yes by existing tests. More tests needed to verify abort behavior may be desired.

## Are there any user-facing changes?

No, just faster writes!